Incorporated translated changes from O2 master branch.
Select is not a Mode anymore, now it's on Essential Roles chapter.

The whole chapter 3 was removed. Instead of a whole "Rights and Duties" thing, now
expectations and authorities are distributed among chapters.

Energization is defined as a concept and used to make text simpler.

Circle can get more autonomous by distributing the Circle Lead role.

Third person "Partners" was only way, now there are a lot of sentences in second person "you"

Generally it's easier to understand.